Improve logical flow editor for deletion with physical flows
#1775
