+++ waltz-ng/client/logical-flow/components/source-and-target-panel/source-and-target-panel.js
@@ -16,9 +16,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import _ from "lodash";
-import {nest} from "d3-collection";
 import {event} from "d3-selection";
 import {initialiseData} from "../../../common";
+import {mkTweakers} from '../source-and-target-graph/source-and-target-utilities';
 
 
 const bindings = {
@@ -106,74 +106,6 @@ function calculateSourceAndTargetFlowsByEntity(primaryEntity, logical = []) {
         sourceFlowsByEntityId,
         targetFlowsByEntityId
     };
-}
-
-
-const iconCodes = {
-    // codes from: http://fontawesome.io/cheatsheet/  (conversion: &#x1234; -> \u1234)
-    files: '\uf0c5',
-    file: '\uf016',
-    question: '\uf128',
-    questionCircle: '\uf29c',
-    folder: '\uf115'
-};
-
-
-function toIcon(count = 0) {
-    switch (count) {
-        case 0:
-            return {
-                code: iconCodes.questionCircle,
-                description: 'No physical files specified',
-                color: '#c66'
-            };
-        case 1:
-            return {
-                code: iconCodes.file,
-                description: 'One linked physical files',
-                color: '#000'
-            };
-        case 2:
-            return {
-                code: iconCodes.files,
-                description: 'Two linked physical files',
-                color: '#000'
-            };
-        default:
-            return {
-                code: iconCodes.folder,
-                description: 'Several linked physical files',
-                color: '#000'
-            };
-    }
-}
-
-
-function mkTweakers(tweakers = {},
-                    physicalFlows = [],
-                    logicalFlows = []) {
-
-    const toIdentifier = (entRef) => `${entRef.kind}/${entRef.id}`;
-
-    const logicalFlowsById = _.keyBy(logicalFlows, 'id');
-
-
-    const countPhysicalFlows = (direction) =>
-        _.countBy(physicalFlows, pf => {
-            const logicalFlow = logicalFlowsById[pf.logicalFlowId];
-            return logicalFlow
-                ? toIdentifier(logicalFlow[direction])
-                : null;
-        });
-
-    const sourceCounts = countPhysicalFlows('source');
-    const targetCounts = countPhysicalFlows('target');
-
-
-    tweakers.source.icon = (appRef) => toIcon(sourceCounts[toIdentifier(appRef)]);
-    tweakers.target.icon = (appRef) => toIcon(targetCounts[toIdentifier(appRef)]);
-
-    return Object.assign({} , tweakers);
 }
